Create MagneticSlider.js
Improved source code, handles deadzone, per-mark scaling of the global magnetic radius, and allows statefulness and onChange hooks to be (optionally) passed from the parent, complying with the @material-ui/core/Slider API spec. Pre-processes the marks passed in the props once with a HoC to sort them by (ASC) value, and clean out our API fields from the props to pass a clean version to the underlying Slider component.